requirements.txt
mistralai>=1.0.0
python-dotenv
deepeval>=1.0.0
pytest>=8.0.0
